fix: check that encKey and data are defined before attempting decrypt
Currently, the `userinfo` method expects to receive a response from the sgid
userinfo endpoint that has the type:

{
  sub: string
  key: string
  data: Record<string, string>
}

However, if you check the sgid-server repo, this shape is not correct: in
particular, the key and data attributes are possibly undefined. This happens
when the client scope is just `openid` (i.e. it doesn't request any data
fields). In such cases, the decrypt payload step fails because it attempts
to decrypt the payload using a key that is undefined.

This commit checks that the encryptedPayloadKey and data are not undefined
before attempting to decrypt the payload.