feat: update ci so that package is published upon merging to `release` branch
Currently, we publish to npm after a release is published on GitHub.
This is problematic because anyone on the sgID team can publish a release - there
are no protections or approval flow surrounding publishing a release.
This is dangerous because a bad actor only needs to compromise one
account to release a new, malicious version of the SDK.

This commit changes that by publishing to npm only upon merging to
the release branch, which should be protected.