feat: add separate method for parsing stringifed array data
This commit reverts the functionality of the `userinfo` method
so that it continues to return string data values, and introduces a
new function, `parseData`, which can help to parse stringified array
or objects.

The reason for making the data parsing step an explicit, separate step
is that if we modify the return type of the `userinfo` method, this
is a potentially breaking change, since existing users expect a string
to be returned as a data value.

In practice, there's no breaking change because no array data fields
are currently available in sgID. However, it would still represent a
change in the types returned for the user. It is also a cleaner interface
to consistently return a string and allow the user to parse it if
necessary.