feat: handle array data values (#64)
* feat: add helper functions for parsing stringified arrays

This commit adds the necessary helper functions that help us to
parse stringified arrays, as well as the associated unit tests.

* feat: parse user data as arrays if the decrypted data is wrapped in square brackets

This commit introduces code that attempts to parse the decrypted data
as an array if it is wrapped in square brackets. In the event that
the data is not valid JSON, we return the string in its entirety.

* feat: add separate method for parsing stringifed array data

This commit reverts the functionality of the `userinfo` method
so that it continues to return string data values, and introduces a
new function, `parseData`, which can help to parse stringified array
or objects.

The reason for making the data parsing step an explicit, separate step
is that if we modify the return type of the `userinfo` method, this
is a potentially breaking change, since existing users expect a string
to be returned as a data value.

In practice, there's no breaking change because no array data fields
are currently available in sgID. However, it would still represent a
change in the types returned for the user. It is also a cleaner interface
to consistently return a string and allow the user to parse it if
necessary.

* chore: make `parseData` function a normal method instead of a static method

This commit modifies the parseData function to be a normal class method
because we use the `SgidClient` name as the convention for the singleton
instance, which makes it confusing for us to use static functions.

* chore: detect all stringified objects instead of just stringified arrays

* chore: rename function name for clarity