feat: expose id token in callback method
This commit exposes the id token in the callback method.

The id token is useful because it acts as proof that the user has
been authenticated by sgID. Some relying parties want to store
the id token as part of their audit trail. This is why we're exposing
the id token here.